career-orchestrator/pyproject.toml
[project]
name = "career-orchestrator"
version = "0.1.0"
description = "AI Career-to-Project Orchestrator - JD matching score + project planning"
readme = "../README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Career Orchestrator Team"}
]
keywords = ["ai", "career", "langgraph", "llm", "resume", "job-description"]

dependencies = [
    "pydantic>=2.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
